+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/v5/dataset/CurrentStateRetriever.java
@@ -1,4 +1,4 @@
-package nl.knaw.huygens.timbuctoo.v5.datastores.rssource;
+package nl.knaw.huygens.timbuctoo.v5.dataset;
 
 import nl.knaw.huygens.timbuctoo.v5.datastores.implementations.bdb.BdbTripleStore;
 import nl.knaw.huygens.timbuctoo.v5.datastores.quadstore.dto.CursorQuad;
