+++ timbuctoo-instancev4/src/test/java/nl/knaw/huygens/timbuctoo/v5/datastores/rssource/ChangesListBuilderTest.java
@@ -1,8 +1,8 @@
 package nl.knaw.huygens.timbuctoo.v5.datastores.rssource;
 
 import com.google.common.collect.Lists;
+import nl.knaw.huygens.timbuctoo.v5.dataset.ChangesRetriever;
 import nl.knaw.huygens.timbuctoo.v5.datastores.implementations.bdb.BdbTruePatchStore;
-import nl.knaw.huygens.timbuctoo.v5.datastores.implementations.bdb.UpdatedPerPatchStore;
 import nl.knaw.huygens.timbuctoo.v5.datastores.quadstore.dto.ChangeType;
 import nl.knaw.huygens.timbuctoo.v5.datastores.quadstore.dto.CursorQuad;
 import nl.knaw.huygens.timbuctoo.v5.datastores.quadstore.dto.Direction;
